+++ library/Emphloyer/AbstractJob.php
@@ -7,6 +7,7 @@ namespace Emphloyer;
  */
 abstract class AbstractJob implements Job {
   protected $attributes = array();
+  protected $type = 'job';
 
   /**
    * Once a job has been enqueued it must have some sort of unique id.
@@ -14,6 +15,24 @@ abstract class AbstractJob implements Job {
    */
   public function getId() {
     return $this->attributes['id'];
+  }
+
+  /**
+   * Return the type of the job. Setting the type in your own implementations
+   * allows you to control how much bandwith particular jobs get.
+   * @return string
+   */
+  public function getType() {
+    return $this->type;
+  }
+
+  /**
+   * Set the type of the job.
+   * @param string $type
+   * @return void
+   */
+  public function setType($type) {
+    $this->type = $type;
   }
 
   /**