+++ tests/Emphloyer/WorkshopTest.php
@@ -7,15 +7,31 @@ class WorkshopTest extends \PHPUnit_Framework_TestCase {
     $this->boss = $this->getMockBuilder('Emphloyer\Boss')
       ->disableOriginalConstructor()
       ->getMock();
-    $this->workshop = new Workshop($this->boss, 2);
+    $this->workshop = new Workshop($this->boss, array(array('employees' => 2)));
   }
 
   public function testConstructor() {
     $boss = new Boss($this->getMockBuilder('Emphloyer\Pipeline')->disableOriginalConstructor()->getMock());
-    $workshop = new Workshop($boss, 2);
+    $workshop = new Workshop($boss, array(array('employees' => 2), array('employees' => 1, 'only' => array('special'))));
 
     $employees = $boss->getEmployees();
-    $this->assertEquals(2, count($employees));
+    $this->assertEquals(3, count($employees));
+
+    $countRegular = 0;
+    $countSpecial = 0;
+    foreach ($employees as $employee) {
+      $this->assertInstanceOf('Emphloyer\Employee', $employee);
+      $options = $employee->getOptions();
+      if (isset($options['only'])) {
+        $this->assertEquals(array('employees' => 1, 'only' => array('special')), $options);
+        $countSpecial += 1;
+      } else {
+        $countRegular += 1;
+      }
+    }
+
+    $this->assertEquals(1, $countSpecial);
+    $this->assertEquals(2, $countRegular);
   }
 
   public function testRun() {